Maj type des coord (double vers int)

+++ src/robotsim/model/ChargingStation.java
@@ -4,7 +4,7 @@ import shapes.Rectangle;
 public class ChargingStation extends Component{
 	private boolean busy;
 	
-	public ChargingStation(double x, double y, int width, int height, boolean busy) {
+	public ChargingStation(int x, int y, int width, int height, boolean busy) {
 		super(x, y, new Rectangle(width, height));
 		this.busy = busy;
 	}
